build.rs: fix with rustc not built from Git
I couldn't build hyperx because my rustc (from Nixpkgs) has a version
output looks like this:

    rustc 1.34.2

There's no Git information in parentheses after the version. I assume
that this is because it was not built from a Git checkout.

To fix this, I made it so that if no space occurs after the "rustc "
prefix, the whole remaining string is used as the version.